niffler_e_2_e_tests_python/pyproject.toml
[project]
name = "test_niffler_project"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "fastapi-pagination (>=0.13.1,<0.14.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "playwright (>=1.52.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pydantic = "^2.11.5"
allure-pytest = "^2.14.2"
allure-python-commons = "^2.14.2"
email-validator = "^2.0.1"
python-dotenv = "^1.0.0"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = '.'
addopts = [
    "--alluredir=allure-results"
]